+++ services/device-registry-mongodb/src/test/java/org/eclipse/hono/deviceregistry/mongodb/service/MongoDbBasedRegistrationServiceTest.java
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -94,7 +94,7 @@ public class MongoDbBasedRegistrationServiceTest implements RegistrationServiceT
 
         registrationService.setAutoProvisioner(autoProvisioner);
 
-        registrationService.start().onComplete(testContext.completing());
+        registrationService.createIndices().onComplete(testContext.completing());
     }
 
     /**